+++ src/com/offerII/L016/NoRepeatLongestStr.java
@@ -31,7 +31,7 @@ import java.util.Map;
  * 0 <= s.length <= 5 * 104
  * s 由英文字母、数字、符号和空格组成
  *
- * 注意：本题与主站 3 题相同： https://leetcode-cn.com/problems/longest-substring-without-repeating-characters/
+ * 注意：本题与主站 3 题相同： <a href="https://leetcode-cn.com/problems/longest-substring-without-repeating-characters/">https://leetcode-cn.com/problems/longest-substring-without-repeating-characters/</a>
  */
 public class NoRepeatLongestStr {
     /**
@@ -57,13 +57,13 @@ public class NoRepeatLongestStr {
             //进行窗口内数据的一系列更新
             window.put(c,window.getOrDefault(c,0)+1);
 
-            while (window.get(c).intValue()>1){
+            while (window.get(c) > 1) {
                 //d是将移除窗口的字符
                 char d = s.charAt(left);
                 //左移动窗口
                 left++;
 
-                window.put(d,window.get(d)-1);
+                window.put(d, window.get(d) - 1);
             }
             //更新答案
             res = Math.max(res,right - left);
